✨feat: 전체적인 폴더 구조 및 로직 수정
