+++ src/app/more/MoreList.styles.ts
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const MoreListLayout = styled.section`
+  padding-top: 6rem;
+`;
+
+export const MoreList = styled.ul`
+  padding: 0 1.5rem;
+`;
