+++ src/app/more/SortTab/SortTab.constants.ts
@@ -1,0 +1,17 @@
+import { SortDataType, SortTabType } from "./SortTab.types";
+
+export const SORT_DATA: SortDataType[] = [
+  {
+    title: "최신순",
+    keyword: "latest",
+  },
+  {
+    title: "인기순",
+    keyword: "popular",
+  },
+];
+
+export const SORT_TAB: SortTabType = {
+  LATEST: "latest",
+  POPULAR: "popular",
+};
