+++ src/app/more/SortTab/SortTab.types.ts
@@ -1,0 +1,9 @@
+export interface SortDataType {
+  title: string;
+  keyword: string;
+}
+
+export interface SortTabType {
+  LATEST: "latest";
+  POPULAR: "popular";
+}
