--- src/app/more/[keyword]/page.tsx
@@ -1,7 +1,0 @@
-import MoreListView from "./MoreList.view";
-
-const More = () => {
-  return <MoreListView />;
-};
-
-export default More;
